Separate code in new file
