+++ frontend/frame-generator-frontend.cabal
@@ -22,6 +22,7 @@ library
       Lib
 
   other-modules:
+      EditForeGround
 
   build-depends:
       base
